refactor controllerregistration
